Move some content from correctness_test.go to utils.go
This change makes these functions/types available from the metrics
package, where they will be needed to address issue #652.
